Closes #11: Reorganize data source extensions.
