+++ setup.py
@@ -22,9 +22,9 @@ setup(
     entry_points={
         'redash.extensions': [
             'dockerflow = redash_stmo.dockerflow:dockerflow',
-            'datasource_health = redash_stmo.health:datasource_health',
-            'datasource_link = redash_stmo.datasource_link:datasource_link',
-            'datasource_version = redash_stmo.datasource_version:datasource_version',
+            'datasource_health = redash_stmo.data_sources.health:health',
+            'datasource_link = redash_stmo.data_sources.link:link',
+            'datasource_version = redash_stmo.data_sources.version:version',
         ],
     },
     classifiers=[
