+++ src/redash_stmo/data_sources/link/bundle/datasource_link.js
