+++ src/redash_stmo/data_sources/version/bundle/datasource_version.js
