+++ tests/test_datasource_link.py
@@ -5,16 +5,16 @@ from flask import Flask
 
 from redash.models import DataSource
 from redash.query_runner.pg import PostgreSQL
-from redash_stmo.datasource_link import datasource_link
+from redash_stmo.data_sources.link import link
 
 
 class TestDatasourceLink(BaseTestCase):
     EXPECTED_DOC_URL = "www.example.com"
     def setUp(self):
         super(TestDatasourceLink, self).setUp()
-        self.patched_query_runners = self._setup_mock('redash_stmo.datasource_link.query_runners')
+        self.patched_query_runners = self._setup_mock('redash_stmo.data_sources.link.query_runners')
         self.patched_query_runners.return_value = {}
-        datasource_link(self.app)
+        link(self.app)
 
     def _setup_mock(self, function_to_patch):
         patcher = mock.patch(function_to_patch)
