+++ tests/test_datasource_version.py
@@ -6,7 +6,7 @@ from flask import Flask
 
 from redash.models import DataSource
 from redash.query_runner.pg import PostgreSQL
-from redash_stmo.datasource_version import datasource_version
+from redash_stmo.data_sources.version import version
 
 
 class TestDatasourceVersion(BaseTestCase):
@@ -17,7 +17,7 @@ class TestDatasourceVersion(BaseTestCase):
         self.data_source = self.factory.create_data_source()
         self.patched_run_query = self._setup_mock('redash.query_runner.pg.PostgreSQL.run_query')
         self.patched_runner_type = self._setup_mock('redash.query_runner.pg.PostgreSQL.type')
-        datasource_version(self.app)
+        version(self.app)
 
     def _setup_mock(self, function_to_patch):
         patcher = mock.patch(function_to_patch)
